requirements.txt
flask-session
gunicorn
openai
flask-cors
psycopg2-binary
bcrypt
